requirements.txt
pip>=22.1.1
python-dotenv==1.0.1
django-webpack-loader >= 3.1.0
psycopg2>=2.9.9
